+++ CHANGELOG.md
@@ -19,6 +19,7 @@ for open development tasks and limitations.
  - pdf 
    - watermark   
 - `scan` BUG-FIX: raise StopIteration removed when file is renamed in repo folder
+- `check` sub-cmd, check db and repo integrity
 -